flaskBackend/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
Flask==3.1.0
flask-cors==5.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
nltk==3.9.1
oauthlib==3.2.2
praw==7.8.1
prawcore==2.4.0
pycparser==2.22
PyJWT==2.10.1
pymongo==4.11.3
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
textblob==0.19.0
tqdm==4.67.1
tweepy==4.15.0
update-checker==0.18.0
urllib3==2.3.0
vaderSentiment==3.3.2
websocket-client==1.8.0
Werkzeug==3.1.3